+++ Source/Plugins/Core/com.equella.core/js/tsrc/modules/CollectionsModule.ts
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 import * as OEQ from "@openequella/rest-api-client";
-import { API_BASE_URL } from "../config";
+import { API_BASE_URL } from "../AppConfig";
 import { listEntities } from "./OEQHelpers";
 
 /**
